Start the render thread with the actual current buffer size
This prevents rendering to places for which no buffer was allocated yet.  Normally, self.{Width,Height} and Buffer.{Width,Height} are equal.  Except during property loading at application start.  Then it is possible for self.{Width,Height} to contain the loaded properties of the form and Buffer.{Width,Height} to still contain the default values (192, 192).  If the render thread is started with a bigger rect than this, it could lead to memory access violations on a sufficiently fast machine or with weird timing exactly when rendering outside (192, 192) is done before the actual Buffer for it is allocated in the main thread.

git-svn-id: svn://svn.code.sf.net/p/graphics32/code/trunk@1367 dadbf757-a53b-4b3f-9c79-7c22a95d7d7a